services/clip-runpod-worker/requirements.txt
# CLIP RunPod Worker Dependencies (Always-On HTTP Service)
# Pinned versions for reproducibility

# FastAPI + ASGI server
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP client for async downloads
aiohttp==3.11.10

# CLIP (pinned - note: torch installed separately in Dockerfile)
open-clip-torch==2.26.1

# NumPy (torch 2.x compatible)
numpy==1.26.4

# Image processing
pillow==11.0.0

# System/utilities
python-multipart==0.0.20
